refactor: Wave 2 - extract client MP construct hooks to Core/Net/
Split the two client-side construct-message hooks (RegisterClientConstructChunkGuardHook +
RegisterClientGridChunkFireHook) and their file-local helpers/statics verbatim out of
SFGameInstanceModule.cpp into Core/Net/SFGameInstanceModule_NetHooks.cpp. They remain
USFGameInstanceModule members, registered identically from StartupModule - no behaviour change.
Verified the moved statics share no state with the module's other hooks (1452 -> 978 lines).

MP-validated on a Windows dedicated server: as a client, scaled/auto-connect builds and Extend
all construct cleanly - the server consumed the staged scaling specs + extend commits (proving
the relocated client fire-hook fired), with 0 failed constructs/serialize and clean chain sweeps.
